feat(app): show generic tool details
